+++ pyArango/database.py
@@ -59,7 +59,7 @@ class Database(object) :
 
                 self.collections[colName] = colObj
         else :
-            raise updateError(data["errorMessage"], data)
+            raise UpdateError(data["errorMessage"], data)
 
     def reloadGraphs(self) :
         "reloads the graph list"
@@ -170,10 +170,12 @@ class Database(object) :
         """returns true if the databse has a graph by the name of 'name'"""
         return name in self.graphs
 
-    def AQLQuery(self, query, batchSize = 100, rawResults = False, bindVars = {}, options = {}, count = False, fullCount = False, **moreArgs) :
+    def AQLQuery(self, query, batchSize = 100, rawResults = False, bindVars = {}, options = {}, count = False, fullCount = False,
+                 json_encoder = None, **moreArgs) :
         """Set rawResults = True if you want the query to return dictionnaries instead of Document objects.
         You can use **moreArgs to pass more arguments supported by the api, such as ttl=60 (time to live)"""
-        return AQLQuery(self, query, rawResults = rawResults, batchSize = batchSize, bindVars  = bindVars, options = options, count = count, fullCount = fullCount, **moreArgs)
+        return AQLQuery(self, query, rawResults = rawResults, batchSize = batchSize, bindVars  = bindVars, options = options, count = count, fullCount = fullCount,
+                        json_encoder = json_encoder, **moreArgs)
 
     def explainAQLQuery(self, query, allPlans = False) :
         """Returns an explanation of the query. Setting allPlans to True will result in ArangoDB returning all possible plans. False returns only the optimal plan"""
@@ -206,7 +208,7 @@ class Database(object) :
 
         data = r.json()
 
-        if r.status_code == 200 and not data["error"] :
+        if (r.status_code == 200 or r.status_code == 201 or r.status_code == 202) and not data["error"] :
             return data
         else :
             raise TransactionError(data["errorMessage"], action, data)