+++ pyArango/document.py
@@ -1,15 +1,177 @@
 import json, types
 
-from .theExceptions import (CreationError, DeletionError, UpdateError)
+from .theExceptions import (CreationError, DeletionError, UpdateError, ValidationError, SchemaViolation, InvalidDocument)
 
-__all__ = ["Document", "Edge"]
+__all__ = ["DocumentStore", "Document", "Edge"]
+
+class DocumentStore(object) :
+    """Store all the data of a document in hierarchy of stores and handles validation.
+    Does not store private information, these are in the document."""
+
+    def __init__(self, collection, validators={}, initDct={}, patch=False, subStore=False) :
+        self.store = {}
+        self.patchStore = {}
+        self.collection = collection
+        self.validators = validators
+        
+        self.mustValidate = False
+        for v in self.collection._validation.values() :
+            if v :
+                self.mustValidate = True
+                break
+        
+        self.isSubStore = subStore
+        self.subStores = {}
+        self.patching = patch
+        self.set(initDct)
+        self.patching = True
+
+    def resetPatch(self) :
+        """reset patches"""
+        self.patchStore = {}
+
+    def getPatches(self) :
+        """get patches as a dictionary"""
+        if not self.mustValidate :
+            return self.getStore()
+
+        res = {}
+        res.update(self.patchStore)
+        for k, v in self.subStores.items() :
+            res[k] = v.getPatches()
+        
+        return res
+        
+    def getStore(self) :
+        """get the inner store as dictionary"""
+        res = {}
+        res.update(self.store)
+        for k, v in self.subStores.items() :
+            res[k] = v.getStore()
+        
+        return res
+
+    def validateField(self, field) :
+        """Validatie a field"""
+        if field not in self.validators and not self.collection._validation['allow_foreign_fields'] :
+            raise SchemaViolation(self.collection.__class__, field)
+
+        if field in self.store:
+            if isinstance(self.store[field], DocumentStore) :
+                return self[field].validate()
+            
+            if field in self.patchStore :
+                return self.validators[field].validate(self.patchStore[field])
+            else :
+                return self.validators[field].validate(self.store[field])
+
+        return True
+
+    def validate(self) :
+        """Validate the hole document"""
+        if not self.mustValidate :
+            return True
+
+        res = {}
+        for field in self.validators.keys() :
+            try :
+                if isinstance(self.validators[field], dict) and field not in self.store :
+                    self.store[field] = DocumentStore(self.collection, validators = self.validators[field], initDct = {}, subStore=True)
+                self.validateField(field)
+            except InvalidDocument as e :
+                res.update(e.errors)
+            except (ValidationError, SchemaViolation) as e:
+                res[field] = str(e)
+
+        if len(res) > 0 :
+            raise InvalidDocument(res)
+        
+        return True
+
+    def set(self, dct) :
+        """Set the store using a dictionary"""
+        if not self.mustValidate :
+            self.store = dct
+            self.patchStore = dct
+            return
+
+        for field, value in dct.items() :
+            if field not in self.collection.arangoPrivates :
+                if isinstance(value, dict) :
+                    if field in self.validators :
+                        vals = self.validators[field]
+                    else :
+                        vals = {}
+                    self[field] = DocumentStore(self.collection, validators = vals, initDct = value, patch = self.patching, subStore=True)
+                    self.subStores[field] = self.store[field]
+                else :
+                    self[field] = value
+
+    def __getitem__(self, field) :
+        """Get an element from the store"""
+        if (field in self.validators) and isinstance(self.validators[field], dict) and (field not in self.store) :
+            self.store[field] = DocumentStore(self.collection, validators = self.validators[field], initDct = {}, patch = self.patching, subStore=True)
+            self.subStores[field] = self.store[field]
+            self.patchStore[field] = self.store[field]
+
+        if self.collection._validation['allow_foreign_fields'] or self.collection.hasField(field) :
+            return self.store.get(field)
+
+        try :
+            return self.store[field]
+        except KeyError :
+            raise SchemaViolation(self.collection.__class__, field)
+
+    def __setitem__(self, field, value) :
+        """Set an element in the store"""
+        if (not self.collection._validation['allow_foreign_fields']) and (field not in self.validators) and (field not in self.collection.arangoPrivates):
+            raise SchemaViolation(self.collection.__class__, field)
+        
+        if field in self.collection.arangoPrivates :
+            raise ValueError("DocumentStore cannot contain private field (got %s)" % field)
+
+        if isinstance(value, dict) :
+            if field in self.validators :
+                vals = self.validators[field]
+            else :
+                vals = {}
+            self.store[field] = DocumentStore(self.collection, validators = vals, initDct = value, patch = self.patching, subStore=True)
+            
+            self.subStores[field] = self.store[field]
+        else :
+            self.store[field] = value
+
+        if self.patching :
+            self.patchStore[field] = self.store[field]
+
+        if self.collection._validation['on_set'] :
+            self.validateField(field)
+
+    def __delitem__(self, k) :
+        """removes an element from the store"""
+        try :
+            del(self.store[k])
+        except :
+            pass
+    
+        try :
+            del(self.patchStore[k])
+        except :
+            pass
+
+    def __contains__(self, k) :
+        """returns true or false weither the store has a key k"""
+        return (k in self.store) or (k in self.validators)
+
+    def __repr__(self) :
+        return "<store: %s>" % repr(self.store)
 
 class Document(object) :
     """The class that represents a document. Documents are meant to be instanciated by collections"""
 
     def __init__(self, collection, jsonFieldInit = {}) :
-        self.reset(collection, jsonFieldInit)
         self.typeName = "ArangoDoc"
+        self.reset(collection, jsonFieldInit)
 
     def reset(self, collection, jsonFieldInit = {}) :
         """replaces the current values in the document by those in jsonFieldInit"""
@@ -17,59 +179,53 @@ class Document(object) :
         self.connection = self.collection.connection
         self.documentsURL = self.collection.documentsURL
 
-        self._store = {}
-        self._patchStore = {}
+        self.setPrivates(jsonFieldInit)
+        self._store = DocumentStore(self.collection, validators=self.collection._fields, initDct=jsonFieldInit)
+        if self.collection._validation['on_load']:
+            self.validate()
 
-        self._id, self._rev, self._key = None, None, None
-        self.URL = None
-
-        self.set(jsonFieldInit)
         self.modified = True
 
+    def validate(self) :
+        """validate the document"""
+        # print self._store
+        self._store.validate()
+        for pField in self.collection.arangoPrivates :
+            self.collection.validatePrivate(pField, getattr(self, pField))
+
     def setPrivates(self, fieldDict) :
-        """will set self._id, self._rev and self._key field. Private fields (starting by '_') are all accessed using the self. interface,
-        other fields are accessed through self[fielName], the same as regular dictionnary in python"""
+        """will set self._id, self._rev and self._key field."""
+        
         try :
-            self._id = fieldDict["_id"]
+            for priv in ["_id", "_key", "_rev"] :
+                setattr(self, priv, fieldDict[priv])
             self.URL = "%s/%s" % (self.documentsURL, self._id)
-            del(fieldDict["_id"])
-
-            self._rev = fieldDict["_rev"]
-            del(fieldDict["_rev"])
-
-            self._key = fieldDict["_key"]
-            del(fieldDict["_key"])
         except KeyError :
-            self._id, self._rev, self._key = None, None, None
+            for priv in ["_id", "_key", "_rev"] :
+                setattr(self, priv, None)
             self.URL = None
 
-    def set(self, fieldDict = None) :
-        """Sets the document according to values contained in the dictinnary fieldDict. This will also set self._id/_rev/_key"""
-
-        if fieldDict and self._id is None :
-            self.setPrivates(fieldDict)
-
-        if self.collection._validation['on_set']:
-            for k in list(fieldDict.keys()) :
-                self[k] = fieldDict[k]
-        else :
-            self._store.update(fieldDict)
+    def set(self, fieldDict) :
+        """set the document with a dictionary"""
+        self._store.set(fieldDict)
 
     def save(self, waitForSync = False, **docArgs) :
         """Saves the document to the database by either performing a POST (for a new document) or a PUT (complete document overwrite).
-        If you want to only update the modified fields use the .path() function.
+        If you want to only update the modified fields use the .patch() function.
         Use docArgs to put things such as 'waitForSync = True' (for a full list cf ArangoDB's doc).
         It will only trigger a saving of the document if it has been modified since the last save. If you want to force the saving you can use forceSave()"""
+        payload = self._store.getStore()
+        self._save(payload, waitForSync = False, **docArgs)
+
+    def _save(self, payload, waitForSync = False, **docArgs) :
 
         if self.modified :
-            if self.collection._validation['on_save'] :
-                self.validate(patch = False)
 
             params = dict(docArgs)
             params.update({'collection': self.collection.name, "waitForSync" : waitForSync })
-            payload = {}
-            payload.update(self._store)
 
+            if self.collection._validation['on_save'] :
+                self.validate()
             if self.URL is None :
                 if self._key is not None :
                     payload["_key"] = self._key
@@ -96,7 +252,7 @@ class Document(object) :
 
             self.modified = False
 
-        self._patchStore = {}
+        self._store.resetPatch()
 
     def forceSave(self, **docArgs) :
         "saves even if the document has not been modified since the last save"
@@ -115,16 +271,18 @@ class Document(object) :
         The default behaviour concening the keepNull parameter is the opposite of ArangoDB's default, Null values won't be ignored
         Use docArgs for things such as waitForSync = True"""
 
-        if self.collection._validation['on_save'] :
-            self.validate(patch = True)
-
         if self.URL is None :
             raise ValueError("Cannot patch a document that was not previously saved")
 
-        if len(self._patchStore) > 0 :
+        payload = self._store.getPatches()
+        
+        if self.collection._validation['on_save'] :
+            self.validate()
+
+        if len(payload) > 0 :
             params = dict(docArgs)
             params.update({'collection': self.collection.name, 'keepNull' : keepNull})
-            payload = json.dumps(self._patchStore)
+            payload = json.dumps(payload)
 
             r = self.connection.session.patch(self.URL, params = params, data = payload)
             data = r.json()
@@ -135,7 +293,7 @@ class Document(object) :
 
             self.modified = False
 
-        self._patchStore = {}
+        self._store.resetPatch()
 
     def delete(self) :
         "deletes the document from the database"
@@ -149,13 +307,6 @@ class Document(object) :
         self.reset(self.collection)
 
         self.modified = True
-
-    def validate(self, patch = False) :
-        "validates either the whole store, or only the patch store( patch = True) of the document according to the collection's settings.If logErrors returns a dictionary of errros per field, else raises exceptions"
-        if patch :
-            return self.collection.validateDct(self._patchStore)
-        else :
-            return self.collection.validateDct(self._store)
 
     def getInEdges(self, edges, rawResults = False) :
         "An alias for getEdges() that returns only the in Edges"
@@ -173,55 +324,59 @@ class Document(object) :
         except AttributeError :
             raise AttributeError("%s does not seem to be a valid Edges object" % edges)
 
-    def __getitem__(self, k) :
-        """Document fields are accessed in a dictionary like fashion: doc[fieldName]. With the exceptions of private fiels (starting with '_')
-        that are accessed as object fields: doc._key"""
-        if self.collection._validation['allow_foreign_fields'] or self.collection.hasField(k) :
-            return self._store.get(k)
+    def getStore(self) :
+        """return the store in a dict format"""
+        return self._store.getStore()
 
-        try :
-            return self._store[k]
-        except KeyError :
-            raise KeyError("Document of collection '%s' has no field '%s', for a permissive behaviour set 'allow_foreign_fields' to True" % (self.collection.name, k))
+    def getPatches(self) :
+        """return the patches in a dict format"""
+        return self._store.getPatches()
+
+    def __getitem__(self, k) :
+        """get an element from the document"""
+        if k in self.collection.arangoPrivates :
+            return getattr(self, k)
+        return self._store[k]
 
     def __setitem__(self, k, v) :
-        """Documents work just like dictionaries doc[fieldName] = value. With the exceptions of private fiels (starting with '_')
-        that are accessed as object fields: doc._key"""
-
-        def _recValidate(k, v) :
-            if type(v) is dict :
-                for kk, vv in v.items() :
-                    newk = "%s.%s" % (k, kk)
-                    _recValidate(newk, vv)
-            else :
-                self.collection.validateField(k, v)
-
-        if self.collection._validation['on_set'] :
-            _recValidate(k, v)
-
-        self._store[k] = v
-        if self.URL is not None :
-            self._patchStore[k] = self._store[k]
-
-        self.modified = True
+        """set an element in the document"""
+        if k in self.collection.arangoPrivates :
+            setattr(self, k, v)
+        else :
+            self._store[k] = v
 
     def __delitem__(self, k) :
+        """removes an element from the document"""
         del(self._store[k])
-
+    
     def __str__(self) :
-        return "%s '%s': %s" % (self.typeName, self._id, repr(self._store))
+        return repr(self)
 
     def __repr__(self) :
-        return "%s '%s': %s" % (self.typeName, self._id, repr(self._store))
+        privStr = []
+        for p in self.collection.arangoPrivates :
+            privStr.append("%s: %s" % (p, getattr(self, p)) )
+
+        privStr = ', '.join(privStr)
+        return "%s '%s': %s" % (self.typeName, privStr, repr(self._store))
 
 class Edge(Document) :
     """An Edge document"""
     def __init__(self, edgeCollection, jsonFieldInit = {}) :
+        self.typeName = "ArangoEdge"
         self.reset(edgeCollection, jsonFieldInit)
 
     def reset(self, edgeCollection, jsonFieldInit = {}) :
         Document.reset(self, edgeCollection, jsonFieldInit)
-        self.typeName = "ArangoEdge"
+
+    def setPrivates(self, fieldDict) :
+        """set _id, _key, _rev, _from, _to"""
+        super(Edge, self).setPrivates(fieldDict)
+        if "_from" in fieldDict :
+            self._from = fieldDict["_from"]
+        
+        if "_to" in fieldDict :
+            self._to = fieldDict["_to"]
 
     def links(self, fromVertice, toVertice, **edgeArgs) :
         """
@@ -233,17 +388,17 @@ class Edge(Document) :
             if not fromVertice._id :
                 fromVertice._id.save()
 
-            self["_from"] = fromVertice._id
+            self._from = fromVertice._id
         elif (type(fromVertice) is bytes) or (type(fromVertice) is str) :
-            self["_from"] = fromVertice
+            self._from  = fromVertice
 
         if toVertice.__class__ is Document :
             if not toVertice._id :
                 toVertice._id.save()
 
-            self["_to"] = toVertice._id
+            self._to = toVertice._id
         elif (type(toVertice) is bytes) or (type(toVertice) is str) :
-            self["_to"] = toVertice
+            self._to = toVertice
 
         self.save(**edgeArgs)
 
@@ -251,12 +406,13 @@ class Edge(Document) :
         """Works like Document's except that you must specify '_from' and '_to' vertices before.
         There's also a links() function especially for first saves."""
 
-        import types
-
-        if "_from" not in self._store or "_to" not in self._store :
+        if not getattr(self, "_from") or not getattr(self, "_to") :
             raise AttributeError("You must specify '_from' and '_to' attributes before saving. You can also use the function 'links()'")
 
-        Document.save(self, **edgeArgs)
+        payload = self._store.getStore()
+        payload["_from"] = self._from
+        payload["_to"] = self._to
+        Document._save(self, payload, **edgeArgs)
 
     def __getattr__(self, k) :
         if k == "_from" or k == "_to" :