+++ pyArango/validation.py
@@ -36,10 +36,14 @@ class Length(Validator) :
         self.maxLen = maxLen
 
     def validate(self, value) :
+        try :
+            length = len(value)
+        except :
+            raise ValidationError("Field '%s' of type '%s' has no length" % (value, type(value)))
+            
         if self.minLen <= len(value) and len(value) <= self.maxLen :
             return True
-
         raise ValidationError("Field must have a length in ['%s';'%s'] got: '%s'" % (self.minLen, self.maxLen, len(value)))
-
+    
     def __str__(self) :
         return "%s[%s, %s]" % (self.__class__.__name__, self.minLen, self.maxLen)